pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nautilus-quants"
version = "0.1.0"
description = "Quantitative trading framework built on Nautilus Trader"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Joe" }
]

dependencies = [
    "nautilus_trader==1.222.0",
    "pandas>=2.3.0",
    "numpy>=2.2.0",
    "pyyaml>=6.0",
    "plotly>=5.18.0",  # For Nautilus native visualization
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/nautilus_quants"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
